backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "faker>=40.11.1",
]
